% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aa_names}
\alias{aa_names}
\title{Names and abbreviations for the 20 standard proteinogenic amino acids plus stop codon.}
\format{A data frame with 21 rows and 3 variables:
\describe{
  \item{Name}{Name of the amino acid or Stop}
  \item{Abr}{Three letter abbreviation of the amino acid or End for Stop}
  \item{AA}{One letter abbreviation of the amino acid or * for Stop}
}}
\usage{
aa_names
}
\description{
A dataset containing the full name, three letter abreviation, and single letter
 abbreviation for the 20 standard proteinogenic amino acids plus Stop.
}
\keyword{datasets}
