% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_hive_codon_tbl.R
\name{build_hive_codon_tbl}
\alias{build_hive_codon_tbl}
\title{Converts an organism specific HIVE Codon Usage Table to a Codon Frequency Table}
\usage{
build_hive_codon_tbl(file, skip = 2, total_codons)
}
\arguments{
\item{file}{the name of the file which the data are to be read from}

\item{skip}{integer, the number of lines of the data file to skip before beginning to read data.}

\item{total_codons}{total number of codons analysed per organism as reported in HIVE Codon Usage Table}
}
\value{
data.frame
}
\description{
For details on the Codon Frequency Table format please see
`reversetrasnalte::check_codon_table()`.
}
