% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_codon_table.R
\name{check_codon_table}
\alias{check_codon_table}
\title{Confirm formatting for user supplied Codon Frequency Table.}
\format{codon_tbl (A Codon Frequency Table):
\describe{
  \item{codon}{Three DNA nucleotides representing the amino acid.}
  \item{aa}{One letter abbreviation of the amino acid or * for Stop}
  \item{prop}{The proportion the specific codon is present in the genome}
}}
\usage{
check_codon_table(codon_tbl)
}
\arguments{
\item{codon_tbl}{A Codon Frequency Table with the following column names,
c("codon", "aa", "prop"). Capitalization of column names should not matter.
The prop column must be of type numeric and all proportions between 0 and 1.
The sum of each codons proportion per amino acid must total roughly to 1
(0.98 to 1.02 allowed).}
}
\value{
An updated Codon Frequency Table with all column names lower case for
downstream use.
}
\description{
Confirm formatting for user supplied Codon Frequency Table.
}
