% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_tbl}
\alias{ecoli_tbl}
\title{An Escherichia coli Codon Frequency Table.}
\format{A data frame with 4 rows and 3 variables:
\describe{
  \item{aa}{One letter abbreviation of the amino acid or * for Stop}
  \item{codon}{Three DNA nucletodies representing the amino acid.}
  \item{prop}{The proportion the specific codon is present in the genome)}
}}
\usage{
ecoli_tbl
}
\description{
Built from HIVE Escherichia coli (562) Codon Usage Table
}
\keyword{datasets}
