% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsapien_tbl}
\alias{hsapien_tbl}
\title{An Homo Sapien Codon Frequency Table.}
\format{A data frame with 4 rows and 3 variables:
\describe{
  \item{aa}{One letter abbreviation of the amino acid or * for Stop}
  \item{codon}{Three DNA nucleotides representing the amino acid.}
  \item{prop}{The proportion the specific codon is present in the genome)}
}}
\usage{
hsapien_tbl
}
\description{
Built from HIVE Homo sapiens (9606) Codon Usage Table
}
\keyword{datasets}
