% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimal_aa_seq}
\alias{minimal_aa_seq}
\title{A minimal amino acid sequence comprised of the made up amino acids X and Y
used for testing and display.}
\format{#' A character vector designed to represent a minimal amino acid
sequence of 10 made up amino acids of X, Y and Z.}
\usage{
minimal_aa_seq
}
\description{
A minimal amino acid sequence comprised of the made up amino acids X and Y
used for testing and display.
}
\keyword{datasets}
