% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nod2_txt}
\alias{nod2_txt}
\title{The amino acid sequence for the human protein Nucleotide-binding oligomerization
domain-containing protein 2 (NOD2).}
\format{#' A character vector of the amino acid sequence for the NOD2 protein
protein sequence comprised of 1040 amino acids.}
\source{
\url{https://www.uniprot.org/uniprot/Q9HC29.fasta}
}
\usage{
nod2_txt
}
\description{
A cannonical amino acid sequence of the human protein Nucleotide-binding
oligomerization domain-containing protein 2 (NOD2), downloaded from the
UniProtKB/Swiss-Prot database.
}
\keyword{datasets}
