% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aa_names}
\alias{aa_names}
\title{Names, abbreviations and codons, for the 20 standard proteinogenic amino acids plus stop codon.}
\format{A data frame with 64 rows and 4 variables:
\describe{
  \item{name}{Name of the amino acid or Stop}
  \item{abr}{Three letter abbreviation of the amino acid or End for Stop}
  \item{aa}{One letter abbreviation of the amino acid or * for Stop}
  \item{codon}{Three nucleotide (A,G,C,T only) codon for the assigned amino acid.}
}}
\usage{
aa_names
}
\description{
A dataset containing the full name, three letter abbreviation, single letter
 abbreviation and codon for the 20 standard proteinogenic amino acids plus Stop.
}
\keyword{datasets}
