% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aa_names}
\alias{aa_names}
\title{Names, abbreviations and codons, for the 20 standard proteinogenic amino acids plus stop codon.}
\format{A data frame with 64 rows and 4 variables:
\describe{
  \item{Name}{Name of the amino acid or Stop}
  \item{Abr}{Three letter abbreviation of the amino acid or End for Stop}
  \item{AA}{One letter abbreviation of the amino acid or * for Stop}
  \item{Codon}{Three nucleotide (A,G,C,T only) codon for the assigned amino acid.}
}}
\usage{
aa_names
}
\description{
A dataset containing the full name, three letter abreviation, single letter
 abbreviationand codon for the 20 standard proteinogenic amino acids plus Stop.
}
\keyword{datasets}
