% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsapien_tbl}
\alias{hsapien_tbl}
\title{A default Codon Frequency Table for translation in homo-sapiens.}
\format{A data frame with 64 rows and 3 variables:
\describe{
  \item{AA}{One letter abbreviation of the amino acid or * for Stop}
  \item{Codon}{Three DNA nucletodies representing the amino acid.}
  \item{AA}{The proportion the specific codon is present in the genome)}
}}
\source{
\url{https://www.kazusa.or.jp/codon/cgi-bin/showcodon.cgi?species=9606}
}
\usage{
hsapien_tbl
}
\description{
A dataset containing the human amino acid single letter abbreviations, each
amino acids respective codon, and the proportion that codon is present in the
indicated homo-sapien genome.
}
\keyword{datasets}
