% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimal_freq_tbl}
\alias{minimal_freq_tbl}
\title{A minimal Codon Frequency Table designed for tests.}
\format{A data frame with 4 rows and 3 variables:
\describe{
  \item{aa}{One letter abbreviation of the amino acid or * for Stop}
  \item{codon}{Three DNA nucletodies representing the amino acid.}
  \item{prop}{The proportion the specific codon is present in the genome)}
}}
\usage{
minimal_freq_tbl
}
\description{
A minimal Codon Frequency Table made up for two test amino acids, called X
and Y which are coded for 3 and 1 codon sequence respectively.
}
\keyword{datasets}
