% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_translate.R
\name{reverse_translate}
\alias{reverse_translate}
\title{Reverse translate amino acid Sequence to DNA nucleotides}
\usage{
reverse_translate(aa, codon_tbl, limit = 0, model)
}
\arguments{
\item{codon_tbl}{A propoerly formated Codon Frequency Table with the following
collumn names, c("Codon", "AA", "Prop"). The Prop column must be of type
numberic and all proportiosn between 0 and 1. The sum of each codons
proportion per amino acid equal roughly 1 (0.98 to 1.02 allowed).}

\item{limit}{Numerical value set to restrict codons to proportions greater
then the set limit. Must be between 0 and 0.49. Defaults to 0, no limit if
not set. When limiting rare codons, the residual proportions are split
equally amongst the remaining codon options.}

\item{model}{Either proportional, equal or gc_bias}

\item{AA}{A character string of amino acids.}
}
\value{
character
}
\description{
Reverse translate amino acid Sequence to DNA nucleotides
}
\examples{

}
